Texture loading and rendering Common3D.hpp ~ use float for texture coordinates
MeshAsset.cpp
~ fix texture coordinates bug

RenderEngine.cpp
- remove debug text print to make Xcode Frame Debugger easier

Puck.hpp
~ assign table texture for now

*/shaders/*
~ change file names to .glsl / .hlsl to make auto syntax highlighting in various editors work

+++ include/RavEngine/Common3D.hpp
@@ -11,7 +11,7 @@ struct Vertex
 
 struct VertexNormalUV : public Vertex{
 	float normal[3];
-	uint16_t uv[2];
+	float uv[2];
 };
 
 struct VertexColor : public Vertex{
